hyper-models/hyper-models/export/hycoclip/pyproject.toml
[project]
name = "hycoclip-onnx-harness"
version = "0.0.0"
description = "HyCoCLIP/MERU → ONNX export harness (hyperbolic model zoo)"
requires-python = ">=3.10"

dependencies = [
  "numpy>=1.26.4,<2.4",
  "pillow>=10",
  "loguru>=0.6",
  "omegaconf>=2.3",
  "torch",
  "torchvision",
  "timm>=0.9",
  "onnxruntime",
  "onnx",
  "onnxscript",
  "huggingface-hub[cli]>=1.3.2",
]
